+++ src/main/java/fr/raksrinana/fallingtree/BlockBreakHandler.java
@@ -2,14 +2,18 @@ package fr.raksrinana.fallingtree;
 
 import fr.raksrinana.fallingtree.tree.Tree;
 import fr.raksrinana.fallingtree.tree.TreeHandler;
+import net.fabricmc.fabric.api.event.player.PlayerBlockBreakEvents;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.text.TranslatableText;
 import net.minecraft.util.Util;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 
-public class BlockBreakHandler{
-	public boolean beforeBlockBreak(World world, PlayerEntity player, BlockPos blockPos){
+public class BlockBreakHandler implements PlayerBlockBreakEvents.Before{
+	@Override
+	public boolean beforeBlockBreak(World world, PlayerEntity player, BlockPos blockPos, BlockState blockState, BlockEntity blockEntity){
 		if(!world.isClient()){
 			if(FallingTreeUtils.isPlayerInRightState(player)){
 				return TreeHandler.getTree(world, blockPos).map(tree -> {