+++ src/main/java/fr/raksrinana/fallingtree/leaves/LeafBreakingHandler.java
@@ -1,13 +1,13 @@
 package fr.raksrinana.fallingtree.leaves;
 
 import io.netty.util.internal.ConcurrentSet;
-import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.world.ServerWorld;
+import net.minecraft.util.math.ChunkPos;
+import net.minecraft.world.chunk.Chunk;
 import java.util.Iterator;
 import java.util.Set;
-import static fr.raksrinana.fallingtree.FallingTreeUtils.isLeafBlock;
 
 public class LeafBreakingHandler implements net.fabricmc.fabric.api.event.lifecycle.v1.ServerTickEvents.EndTick{
 	public static final Set<LeafBreakingSchedule> scheduledLeavesBreaking = new ConcurrentSet<>();
@@ -19,19 +19,16 @@ public class LeafBreakingHandler implements net.fabricmc.fabric.api.event.lifecy
 			LeafBreakingSchedule leafBreakingSchedule = leavesBreak.next();
 			ServerWorld world = leafBreakingSchedule.getWorld();
 			if(leafBreakingSchedule.getRemainingTicks() <= 0){
-				if(world.isChunkLoaded(leafBreakingSchedule.getBlockPos())){
+				Chunk chunk = world.getChunk(leafBreakingSchedule.getBlockPos());
+				ChunkPos chunkPos = chunk.getPos();
+				if(world.isChunkLoaded(chunkPos.x, chunkPos.z)){
 					BlockState state = world.getBlockState(leafBreakingSchedule.getBlockPos());
-					Block block = state.getBlock();
-					if(isLeafBlock(block)){
-						block.randomTick(state, world, leafBreakingSchedule.getBlockPos(), world.getRandom());
-					}
-					else{
-						leavesBreak.remove();
+					state.scheduledTick(world, leafBreakingSchedule.getBlockPos(), world.getRandom());
+					if(state.hasRandomTicks()){
+						state.randomTick(world, leafBreakingSchedule.getBlockPos(), world.getRandom());
 					}
 				}
-				else{
-					leavesBreak.remove();
-				}
+				leavesBreak.remove();
 			}
 			else{
 				leafBreakingSchedule.tick();